requirements.txt
aiohttp==3.11.14
alembic==1.15.1
bcrypt==4.2.1
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
gTTS==2.5.4
gunicorn==23.0.0
pdf2image==1.17.0
pdfminer.six==20231228
PyPDF2==3.0.1
pytesseract==0.3.13
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.38
Werkzeug==3.1.3
pillow==11.1.0
